requirements.txt
requests>=2.31.0
transformers>=4.35.0
torch>=2.1.0
gitpython>=3.1.40
PyGithub>=2.1.1
